split out deploying ECR images, for pipeline
